Some refactoring of UsdSchemaRegistry in preparation for allowing single apply API schemas to include other API schemas.
Separated the multiple apply API schema prim definitions from the single
apply API schema prim definitions in the schema registry. This is helpful
because we use them differently and will eventually populate them
differently. The are property namespace prefixes for multiple apply API
schemas are now stored alongside their prim definitions in the same map.

Moved _FindAndAddPluginSchema into a private helper class that is declared
in the .h and defined in the .cpp. As I make more changes to the prim
definition initialization, this helper class will be useful for organizing helper
functions that act on the registry during initialization without the
implementation details creeping out into the public header.

(Internal change: 2178362)