Refactoring some of the functionality in UsdPrimDefinition for how we build prim definitions.
- The stored SdfPrimSpecHandle in the prim definition has been replaced with an SdfPath
to the prim spec. The prim spec was only being used to get the documentation string for
API schema prim definitions which don't store a mapping to prim metadata. The
documentation can be retrieve via just the path and the path is more useful information
to store anyway during prim definition population.

- The prim definition constructor that took a prim spec has been replaced with a constructor
that takes an SdfPath and it no longer populates the properties. It does, however, map the
prim metadata path for non-API schemas. Prim specs that provide a prim definition's
properties are now always composed in directly through the _ComposePropertiesFrom functions.
_SetPrimSpec has been removed as well for this reason.

- All functions for composing properties are now 'compose unders', i.e. duplicate properties
are not overwritten. This is simpler and more consistent for all the prim definition composition
the happens in the schema registry.

UsdSchemaRegistry has been updated to handle these changes. There's also one minor
functional change in that API schema names that can't be mapped to a valid prim definition
are no longer included in the composed API schema list of a UsdPrimDefinition that is
built with them.

(Internal change: 2178468)