Rework the contract between SdrShaderNode and SdrShaderProperty
The SdrShaderNode::_PostProcessProperties() is potentially modifying internal variables on SdrShaderProperty. To make this less arbitrary we introduce a helper function _ConvertToVStruct() and also a _FinalizeProperty() method on SdrShaderProperty.

We move the _ConformDefaultValue call from the constructor to the new _FinalizeProperty method, such that this conforming procedure can react to any changes to the property that _PostProcessProperties might have made.

Also, simplifying the loop(s) _PostProcessProperties, since we do not need to differentiate between input and output properties.

(Internal change: 2179107)