Add support for different Sdr to Sdf type mappings
Introduces the sdrUsdEncodingVersion metadata tag on shader nodes, which is assumed to be an integer value to select an encoding scheme of shader properties in USD.

If a valid tag is present on a shader node, the encoding version is set on each property of that shader, since the choice for the Sdf type is done on each SdrShaderProperty in the GetTypeAsSdfType method.

The actual encoding is chosen within the _GetTypeAsSdfType function, which now selects the suitable function to determine the encoding. The scheme is clearly setup to add encoding conventions in the future or to retire older ones. The goal is isolate these encoding choices in exactly one place.

Improved _ConformDefaultValue to not crash if a float array of length 2,3,4 has an invalid default value. Also modified the return type to match the Sdf type exactly in case we had a mismatch and could not convert it. We drop the original defaultValue and go with the default value of the type.

Reworks and extends the test:
* The mock OSL parser plugin is setting up properties for various types
* The mock discovery plugin is tagging nodes with the new `sdrUsdEncodingVersion` metadatum. One valid and one invalid
* The testSdrRegistry test is now initializing the plugins once and used it in the existing test function as well as in the new test_UsdEncoding function
* The new test function checks the Sdr-to-Sdf type mapping and especially the difference in behavior when a shader node has been tagged with `sdrUsdEncodingVersion`

(Internal change: 2179895)