Rename UsdGeomPurposeVisibilityAPI to UsdGeomVisibilityAPI.
We realized that it was a mistake to introduce a new API schema
that is specific to purpose visibility, since what we want ultimately
is to be able to author *any* visibility-related opinion on any prim,
not just imageable prims. We're still working out the precise staging
to get there, but for sure, we don't want to have to update the name
of the schema.

(Internal change: 2184429)