[hd] Temporarily disable scene index emulation.
Scene index emulation is a significant new imaging feature that was enabled by changing a config flag in CL 2185534.  This CL is turning off emulation by reverting the config flag, for the sake of stability.  However, the plan is to re-enable the config flag as soon as we have a few performance fixes in.

(Internal change: 2186347)