HasConnectableAPI() now does a full search.
Previously this function (on UsdShadeConnectableAPI) would only check a
cache.  If the type hadn't been cached already then the function would
return false even if it would've been true after caching.  Now we do
the full search algorithm from GetBehavior() to fill the cache if
necessary so clients are no longer on the hook to call GetBehavior()
(indirectly) to get the right results.

The change moves the code from GetBehavior() into another function and
has GetBehavior(), GetBehaviorForType() (which HasConnectableAPI()
calls), and GetBehaviorForPrimTypeId() call the new function.

Please note that this makes GetBehaviorForPrimTypeId() and
GetBehaviorForType() significantly more expensive if the result isn't
cached.  This is necessary to get the correct answer.

testUsdShadeConnectableAPIBehavior was changed to reflect the new
behavior.  One part of a test became redundant and was removed while
parts of two other tests that depend on the new behavior were enabled.
A type lookup bug was also fixed;  this bug prevented a couple of
tests from checking what they were intended to check.

(Internal change: 2187449)