[Hd] PrimDataSourceOverlayCache: don't use a per-prim TBB datastructure.
PrimDataSourceOverlayCache holds, per-prim, a series of attributes that have been computed and added to that prim.  Those new attributes are currently living inside a tbb::concurrent_unordered_map, but it turns out this is super expensive.  It's also not needed in current uses of the PrimDataSourceOverlayCache, since hydra has a (mostly) strong guarantee of only one thread accessing each prim.

We'll need to revisit this if we add hierarchical computations to PrimDataSourceOverlayCache, but at that point we have many solutions that aren't as expensive as a TBB datastructure.

(Internal change: 2188421)