[hd/usdImaging] Fix emulation handling of Get() for cameras.
In HdSceneIndexAdapterSceneDelegate::Get():
1.) If prim type is camera, forward to GetCameraParamValue().
2.) Only try to resolve as a primvar if the prim datasource has primvars.
3.) Move the fallback handling out of _GetPrimvar and into the main function.

In UsdImagingDelegate: don't assert if the renderer asks for primvars on a prim that doesn't have any.  The assert is a relic of a time when everything went through the value cache.

This fixes a case where a custom render delegate was calling sceneDelegate->Get("/camera", "foo"), and it was ending up in the scene index adapter's "primvars" fallback, which was causing UsdImaging to assert because cameras don't define primvars.

(Internal change: 2193717)