[Glf,HdSt] Initial support for built-in barycentric coords
This change allows use of the NV_fragment_shader_barycentric
extension when available. This is disabled by default for
consistency across our currently tested platforms, not all of
which support this feature.

This is currently exposed via GlfContextCaps for consistency
with other GL features, but we plan to move this query down to
Hgi to allow use across other Hgi backends.

Contribution: Thor Hjalmarsson

(Internal change: 2191422)