Allow a typeless prim to query for Connectability.
Previously throwing a coding error when a typeless prim was asked about its connectableAPIBehavior, which is both restrictive and incorrect, since an APISchema could still provide connectability to such a typeless prim.

Added a test to cover the same, where a typeless prim is checked for connectableAPI and then a connectableAPIBehavior bearing APISchema is applied to impart the behavior to this typeless prim.

- Also fixed another bug where certain type's were getting incorrect PrimTypeId created, as their typeName was being queried from UsdSchemaRegistry instead of TfType directly.

(Internal change: 2193009)