[usdImagingGL] When drawing unloaded prims as bounds, fix the transform if the bounds drawmode is applied to an instanced prim.
For USD instance prims, the draw mode adapter assumes that the transform will be provided by the instance adapter.  However, the display-unloaded-as-bounds feature doesn't use the instance adapter, so we need to handle transform for that case directly in the draw mode adapter.

Thanks to Mark Tucker for the solution, proposed in issue #1601.

Fixes #1601

(Internal change: 2193209)