[HgiGL,HgiMetal,HgiVulkan] Added macros to help with declartions
HgiMetal shading language programs are organized such that
most shader functions are members of program scope classes.
Because of this, most shader function pre-declarations are not
allowed. For shader source code which is intended to be used
by GLSL as well as MSL we provide macros which can be used
to wrap such function pre-declarations.

Contribution: Jason Fielder

(Internal change: 2194702)