[hdx] Fix up error-handling logic in HdxSimpleLightTask.
This is a follow-up to change 2199472, which fixed an issue where we would sometimes add uninitialized lights to the lighting context.

This change fixes similar control flow issues, namely: 1.) If a light doesn't exist in the scene, instead of adding a default-constructed light to the lighting context, skip the light. 2.) If a light has invalid shadow parameters, keep using the light in the scene instead of discarding it (with shadows disabled, though).

(Internal change: 2202145)