Make "stencil conditioning" consistent between widget and non-widget render passes in HdxPickTask.
"Stencil conditioning" in HdxPickTask invokes a callback supplied by the host
app to write a stencil into the stencil buffer used when rendering pick ID
AOVs. This stencil buffer is retained when drawing pickables and lets us log
which pickables passed the stencil.

When we added the special drawn-in-front render pass for widgets, we didn't
give any special consideration to conditioning the stencil buffer for drawing
widgets, so widgets could erroneously pass the stencil when they shouldn't have.

Now, we provide a stencil buffer where we had none by using the name
"widgetDepthStencil" which is recognized by name and given a format that
includes bits for our stencil, just like with "depthStencil".  Also, we
separately condition the special widget depthStencil buffer by invoking "stencil
conditioning" on it directly.  This makes picking behavior consistent between
widgets and non-widgets.

(Internal change: 2202527)