[hdPrman] removing PXR_MATERIALX_STDLIB_DIR from matfiltMaterialX.
The PXR_MATERIALX_STDLIB_DIR envar was removed recently to allow for runtime
adjustment of search paths. These search paths include the stdlib directory which
is the parent of the OSL directly we need to include when compiling the generated
MaterialX OSL code. Since we can now have multiple search paths, we are only
adding the stdlib/osl path suffix to the search path if it exists, otherwise we directly
include the path, as is.

(Internal change: 2203092)