hdPrman: Moving code to manage render view, target, outputs and display into render view context. The new code can also create several displays. That is, in one render, we can produce several images, e.g., one containing color and the other normals.
Note that this showed an incompatibility between the UsdRender schema and RenderMan.
The output file name is specified on the render product in the Usd schema and as name of the display in RenderMan.
Hence, it is natural to match render products to displays in RenderMan.
The camera is per-render view and not per-display in RenderMan. However, a UsdRenderProduct inherits from RenderSettingsBase so that there is a per-product camera.

(Internal change: 2204238)