Cache usdz archive traversal for subsequent lookup/queries on external references.
The idea for this change and the initial implementation were
contributed by @marsupial in PR #1578. I did some further
simplifications, cleanup to conform to our standard style,
and added a C++ unit test to exercise iterator behavior.

From the original submission:

While linearly traversing the **usdz** archive, cache the path and header
information so that subsequent lookups can be accelerated.

This brings down load time of a **usdz** file that was taking ~3 minutes
to ~16 seconds, and can often wind up performing better than parsing the
original **usdc** file that generated the **usdz** (likely because IO is
avoided on the references as the files are _in memory_ from the archive).

Fixes #1577

(Internal change: 2205950)