[usdImaging] Fix some native instancer resync bugs.
Fix resync of an instanced gprim bound to a non-instanced material, when the material is resynced: in UsdImagingDelegate::_Resync, we will promote a resync of a USD instance prototype (i.e. a prim in a prototype root) to a resync of the instance; but if an instance prototype depends on a different USD prim (outside of the instance subtree) which gets resynced, we can end up resyncing the instance prototype, e.g. calling InstanceAdapter::ProcessPrimResync(/instance1.proto0_cube).  The simplest way to handle this correctly is just to promote that to a resync of /instance1, which will wipe out/recreate all of the hydra data associated with that USD instance's prototype.

Fix prims being added to/removed from a USD prototype root.  Previously, the hydra prim for a native instancer would register dependencies to the USD prims corresponding to child gprims or child instances.  However, this doesn't pick up prims added/removed from the prototype root.  Now, the native instancer registers a USD dependency on the prototype root.

(Internal change: 2209404)