Avoid layer edits in UsdUtilsComputeAllDependencies
This function uses the _AssetLocalizer helper class in a
mode that should avoid asset path remappings that require
layers to be updated. However, _AssetLocalizer was still
specifying a remapping callback function to the underlying
_FileAnalyzer objects just to record the dependencies in
each file. Even though the function would not change any
asset paths, the presence of the remapping function would
cause _FileAnalyzer to go down a codepath that would
wind up attempting to make layer edits.

_AssetLocalizer now uses _FileAnalyzer's "process asset
path" callback to record dependencies instead of doing
so as a side-effect of the remapping callback. This
allows _AssetLocalizer to pass an invalid callback when
no remapping is needed, avoiding the layer edit codepaths.

(Internal change: 2210835)