Correct and clarify docs for Pcp_FindStartingNodeOfClassHierarchy
These docs had incorrect values for the namespace depths
of various nodes: they indicated the number of path components
in the node's path instead of the parent node's path, which
is what PcpNodeRef::GetNamespaceDepth returns.

Also clarified some of the docs in response to questions
from the code review for change 2210939.

(Internal change: 2211383)