Factor out SdfAssetPath-processing logic so that it can (in followup changes) be shared by materials/lights (as it already is), and now also primvars.  A future change will unify lightColorMap assets also.  For now, also not upgrading RenderSetting params, but we could.
Adding asset support for primvars makes primvar-substitutions in
filename parameters much more robust, and prevents "unhandled
value type" warning spewage.

Noting for the future, also, that the IsSupportedImageFile() check
may need to be bypassed in the future if the asset path is fully
specified by a primvar substitution, as even the filename extension
will not be available in that case at the time of asset-value processing.

(Internal change: 2211387)