Update hdPrman display driver to correctly normalize render outputs based on accumulation rule and type
This change mirrors recent changes to the renderman hdPrman display driver. Aovs that use the min, max, zmin or zmax accumulation rules and are of type float should not be normalized. Aovs that may be set up this way include depth and ids, however ids are likely to be int in practice.

Tested with usdview.

(Internal change: 2212676)