[usdMtlx, hdMtlx] Adding support for multi-output MaterialX nodes.
UsdMtlxReader was only storing one output per node. HdMtlx was
not catching the multi-output nodes, and was therefore setting
the wrong type and missing the needed output parameter on the
connected inputs.

Updating the MaterialX matfilt in HdPrman to have shaders per node
instead of per-output, so that all outputs for multi-output nodes are
contained in one generated Osl Shader and corresponding sdrNode.

Fixes #1581

(Internal change: 2215558)