Improve "widget" drawing in hdSt.
Let hdSt points drawItems be optionally affected by
surface shader code.  Since the widget blending style is
implemented as a surface shader mixin in widgetShader.glslfx,
points, which are not subject to surface shading, would
previously not participate in the widget blending style.
This was visually jarring when working with widgets with
visible points, as the points would garishly draw as if in front of
occluders rather than blending behind like the curves did. Now,
basisCurves and meshes may set the HdDisplayStyle member
"pointsShadingEnabled" to indicate their points should be
shaded as surfaces instead of with constant shading.

(Internal change: 2216338)