Adding UsdBakeMaterialX script to bake MaterialX materials.
The script will output an updated USD file that contains an
additional material variant corresponding to the baked version
of each MaterialX material in an input USD file. The baked
materials are generated using MaterialX::TextureBaker and
will generate a mtlx file and associated texture(s).

Note that this will not be built on Windows until the default
MaterialX version used is at least v1.38.3 where shared
libraries can be built on Windows.

(Internal change: 2216555)