[HdSt,HgiGL] Generate GLSL version and extension declarations using Hgi
The GLSL #version declaration is already being generated using
Hgi and with this change we generate the #extension declarations
using Hgi as well.

As part of this change split out the final compilation step
into a separate method _CompileWithGeneratedGLSLResources().
This separation is useful as we transition to compiling with
full Hgi based generation of shader resources.

Contribution: Jon Creighton, David Yu

(Internal change: 2217154)