Update usdGenSchema to include an elective option to be more acceptable of the authored schema.usda names/values
If useLiteraldentifier metadata is provided for the schema library in the GLOBAL prim, then
usdGenSchema will try to respect authored Property names, attribute default values and allowedTokens as-is,
instead of camelCasing these, which is still the default behavior. Barring scenarios listed below:

- strings beginning with a numeric digit will be prefixed with a '_'
- strings containing ':', will be camelCased [Only for property names, to respect namespace prefixes]
- illegalIdentifier strings will be updated using Tf.MakeValidIdentifier

Also note that prefixing of strings beginning with a numeric digit with '_' will be done unconditionally!
Also if any identifier is illegal, it will be unconfitionally made legal, using Tf.MakeValidIdentifier.

This is helpful in scenarios where schema.usda generation itself is not hand authored but is constructed using usdgenschemafromsdr utility.

(Internal change: 2219026)