[USD/cmake] Reverted disabling Metal on macOS versions older than 10.15
We still want viewport interop to work for Embree and Prman.

Instead of disabling Metal support, we restrict which version of the
Metal Shading Language is used for earlier releases. Specifically,
this allows Metal interop to continue to work for macOS 10.14 (Mojave)
using MTLLanguageVersion2_1.

We recommend building using the latest supported version of Xcode
for the target OS release, e.g. Xcode 11.3.x for macOS 10.14,
Xcode 12.4.x for macOS 10.15, etc.

(Internal change: 2221150)