[HgiMetal] Added parallel encoding of render commands
This is an important optimization for Metal since, in the
absence of support for Multi Draw Indirect (MDI), each
draw command in a compiled indirect draw command buffer
must be encoded separately. Eventually, we'd like to
consider using GPU encoded Indirect Command Buffers (ICB).

An autorelease pool is used to manage the lifetime of
command encoders created within each frame. On Metal
before 12.0 there is a limitation in the Metal Framework
and this pool prevents Command Buffer creation from
leaking and locking up the application.

Also, on macOS prior to 12.2, the Vega GPU driver has a
state leakage bug which causes a GPU hang when drawing
mixed primitive types, such as points and triangles.
This implements a workaround which issues an empty draw
command when the primitive type changes.

Contribution: Jon Creighton, James McCarthy

(Internal change: 2222570)