[HdSt,Hdx] Added GLSLFX resource layouts to shader source
This extends several of the authored glslfx shader source files
to include GLSLFX resource layout dictionaries. For now, we are
adding these only where necessary to support Storm running on
both Metal and GL. Eventually, this system might be used to
further simplify setup of compute shaders and other shaders.

These dictionaries can be parsed by HdSt_ResourceLayout to
extract state that will be used by HdSt_CodeGen.

For now, there is no change in behavior and the existing
resource layout declarations remain guarded by a preprocessor
symbol which will be switched on in a subsequent change.

(Internal change: 2222571)