[HdSt] Initial support for codeGen using GLSLFX resource layouts
This is a partial refactoring of two aspects of codeGen:
1) Gathering shader resources from GLSLFX layouts
2) Generating shader resources using Hgi

Resources are collected into separate buckets similar
to the stringstream generated source code buckets.

The set of resources gathered from the GLSLFX source
are only a subset of the resources that are needed by
a generated shader program. Additional resources, e.g.
attributes, textures, buffers, etc will be added to these
collections in subsequent changes.

This is enabled by default only for Metal.

(Internal change: 2224043)