[HdSt] Fixed some unresolved shader symbols for Metal
The Metal Shading Language program linker requires all refereneced
functions to be defined even if not called. This adds definitions for
"integrateLights()" for image shaders and volumes since that symbol
is referenced from the lighting and lighting termal shader code.

Also, moved the definition of the "GetPatchCoord()" out of general
codeGen an into the mesh shaders specificially, since it is needed
only by mesh primitives.

Contribution: Jon Creighton

(Internal change: 2224562)