[HgiMetal] Enabled Storm for macOS using HgiMetal
The implementation of Storm for Metal requires support for
built-in fragment shader barycentric coords, e.g. we need
support for the [[barycentric_coord]] fragment function
input attribute. So, this is enabled only for macOS >= 10.15
and only on systems with display driver
support for barycentric coords.

Contribution: Jon Creighton

(Internal change: 2225274)