Fix to how we compose reference and payload list ops when layers and  sublayers have layer offsets (both explicitly authored or implicit through  time codes per second resolution).
This easiest way to explain this is with an example. The issue from the
 bug this is for is someone has a root layer with two sublayers:
-root.usda-
(
    subLayers = [
        @./sublayer_del_ref_30tcps.usda@,
        @./sublayer_add_ref.usda@
    ]
)

-sublayer_add_ref.usda-
def Xform "World" (
    prepend references = @./mesh.usda@</World>
) {
}

-sublayer_del_ref_30tcps.usda-
(
    timeCodesPerSecond = 30
)
def Xform "World" (
    delete references = @./mesh.usda@</World>
) {
}

The user is expecting the stronger sublayer to remove the reference to
mesh in /World, but because of the 30 tcps specified in the deleting
layer we resolve a layer offset for the reference to convert it into the 24
tcps of the root layer. This resolved layer changes the identity of deleted
reference and since it no longer matches reference added by the weaker
sublayer, it is not actually deleted.

While the automatic scaling from the tcps difference makes this extra
mysterious, this same problem would be caused if there were no tcps
difference but the sublayers were authored with different layer offsets in
the root layer, e.g.
(
    subLayers = [
        @./sublayer_del_ref_.usda@ (scale = 2.0),
        @./sublayer_add_ref.usda@
    ]
)

This fix is pretty simple in that while we still fully resolve the layer offsets
for the reference arc in the exact same way as before, we only use the
layer offset as authored as part of the identity of the reference when
composing the list op. Now in both of the above examples, the reference
 is properly removed regardless of the offsets of the sublayers.

Fixes #1778

(Internal change: 2226344)