defines HdGpGenerativeProcedural
This is a base class for procedurals which have read access to the input scene and
can generate (or update) prims beneath a given prim. Also includes a
registry for delivering these procedurals via plug-in.

HdGpGenerativeProceduralResolvingSceneIndex can be instantiated to resolve/evaluate/update
these procedurals when found within the input scene index.

(Internal change: 2230703)