[usdImaging] Add edit processing support for UsdImagingStageSceneIndex.
This change subscribes the scene index to the _OnObjectsChanged notice, and turns the notice into a series of "resync" and "invalidate" operations.  Invalidation operations are dispatched through prim adapters, which are responsible for saying (for that prim type) what hydra properties changed for a given set of USD properties.

(Internal change: 2231145)