arch: Add API indicating when the app is crashing
It can be useful for non-crashing threads to know that another thread has
crashed. We now have ArchIsAppCrashing() for this purpose.

Internally we use std::sig_atomic_t to ensure safety with regard to
asynchronous signals. This doesn't guarantee thread safety by itself, but
the only place where the crashing flag is set is protected by a spin lock
that only lets one thread execute at a time, so we end up with a
thread+signal safe mechanism.

(Internal change: 2233779)