[hdSt] Transform two channel (grayscale + alpha) PNG images into RGBA format
Updating the textureChannel test in usdImagingGLBasicDrawing.
The previous textureChannels2.png image had the pattern stored
in the alpha channel, it is now swapped so the pattern is in the R
channel to be more consistant with the other test case pngs.

Note that this test case is just taking the first n channels based on
the original number of channels in the image. For the two channel
case this means we are expanding (r, a) to (r, r, r, a) which this
test case turns it into (r, r, 0, 1)

Added a test case to usdImagingGLBasicDrawing that has an 8 bit
two channel transparent grayscale image.

Fixes #1796

(Internal change: 2235970)