Disable imaging tests on static build configurations
- hdStorm, hdEmbree, etc are all plugins and hdSt sets the default renderer by inspecting the plugin registry.
On static library we do not support building any of the plugin mechanisms (and python).

Above was also the reason for imaging tests failing on static builds -- as the tests were falling back to use legacy renderer as it was not able to find and load any of the renderer plugins.

(Internal change: 2238452)