Avoid opening all value clips when looking for time samples in a given interval.
UsdAttribute::GetTimeSamplesInInterval was previously
implemented by getting all time samples for the attribute
and then copying out just the ones in the specified
interval. If an attribute's time samples came from value
clips, this required all of the clips to be opened.

We now look only at the value clips that are active in
the given interval to avoid this unnecssary cost. The
one exception to this is when interpolation of missing
clip values is enabled and some (or all) clips do not
have authored time samples. In this case, we may look
clips outside the interval to determine if they have
samples. This is a documented limitation of the
interpolation behavior and can be mitigated by
authoring value blocks in the clip manifest.

(Internal change: 2238793)