Fixing an issue in usdview where using the F hotkey to frame the current selection in the prim browser would instead select the first item starting with F if such an item was present in the currently expanded tree widget.
The default behavior of QTreeWidget has been to highlight the first
found item starting with the pressed letter in the current tree
expansion when a letter key is pressed. It seems that starting in
Qt 5.15 though, this behavior was updated to actually select the item
instead of just highlighting it. This change in Qt is what caused the F
hotkey to start changing the selection before framing the selection when
an item starting with F is visible.

The solution is to disable this keyboardSearch functionality for the
widget entirely as we don't want F to ever change the selection and it
doesn't make sense to still allow other letter keys to change the
current selection.

(Internal change: 2239935)