Created separate signal for when camera settings change.
This new signal is a result of refactoring signalFrustumChanged. Previously
signalFrustumChanged was a broad-purpose signal who's handler did more work
than was necessary for some free camera changes. This change adds in a second
signal (emitted when frustum changes aren't related to camera movements) and
splits the old handler into two, updating both so that the appropriate amount of
work is done by each.

(Internal change: 2240544)